+++ src/andFilter.ts
@@ -12,4 +12,8 @@ export class AndFilter implements IFilter {
   public isEquivalentTo(other: IFilter): boolean {
     return areFiltersEquivalent(this, other);
   }
+
+  public get complexity(): number {
+    return this.filters.reduce((sum, f) => sum + f.complexity, 0) * 100 + 10;
+  }
 }