+++ src/startsWithFilter.ts
@@ -15,4 +15,8 @@ export class StartsWithFilter implements IFilter {
   public isEquivalentTo(other: IFilter): boolean {
     return areFiltersEquivalent(this, other);
   }
+
+  public get complexity(): number {
+    return 3;
+  }
 }