+++ test/containsFilter.test.ts
@@ -98,4 +98,22 @@ describe("ContainsFilter", () => {
     // Assert
     expect(result).toBeFalsy();
   });
+
+  describe("complexity", () => {
+    it("should be 1", () => {
+      // Arrange
+      const f = new ContainsFilter("name", "bat");
+
+      // Act / Assert
+      expect(f.complexity).toBe(2);
+    });
+
+    it("should never be 0", () => {
+      // Arrange
+      const f = new ContainsFilter("name", "bat");
+
+      // Act / Assert
+      expect(f.complexity).toBeGreaterThan(0);
+    });
+  });
 });