+++ test/endsWithFilter.test.ts
@@ -54,4 +54,22 @@ describe("EndsWithFilter", () => {
     // Assert
     expect(result).toBeFalsy();
   });
+
+  describe("complexity", () => {
+    it("should be 1", () => {
+      // Arrange
+      const f = new EndsWithFilter("name", "man");
+
+      // Act / Assert
+      expect(f.complexity).toBe(4);
+    });
+
+    it("should never be 0", () => {
+      // Arrange
+      const f = new EndsWithFilter("name", "man");
+
+      // Act / Assert
+      expect(f.complexity).toBeGreaterThan(0);
+    });
+  });
 });