+++ test/startsWithFilter.test.ts
@@ -54,4 +54,22 @@ describe("StartsWithFilter", () => {
     // Assert
     expect(result).toBeFalsy();
   });
+
+  describe("complexity", () => {
+    it("should be 1", () => {
+      // Arrange
+      const f = new StartsWithFilter("name", "Bat");
+
+      // Act / Assert
+      expect(f.complexity).toBe(3);
+    });
+
+    it("should never be 0", () => {
+      // Arrange
+      const f = new StartsWithFilter("name", "Bat");
+
+      // Act / Assert
+      expect(f.complexity).toBeGreaterThan(0);
+    });
+  });
 });